% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcc_correlations.R
\name{cdcc_correlations}
\alias{cdcc_correlations}
\title{This function get the correlation matrix (Rt) of estimated cDCC-GARCH model.}
\usage{
cdcc_correlations(param, stdresids, uncR, ts)
}
\arguments{
\item{param}{cDCC-GARCH parameters(alpha,beta)}

\item{stdresids}{matrix of standrdized(De-GARCH) residual returns  (T by N)}

\item{uncR}{unconditional correlation matrix of stdresids  (N by N)}

\item{ts}{ts how many time series are you taking}
}
\value{
the correlation matrix (Rt) of estimated cDCC-GARCH model (T by N^2)
}
\description{
This function get the correlation matrix (Rt) of estimated cDCC-GARCH model.
}
\note{
Rt are vectorized values of the conditional correlation matrix(Rt) until time t(ts) for each row.
}
