% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcc_loglikelihood.R
\name{cdcc_loglikelihood}
\alias{cdcc_loglikelihood}
\title{This function calculates log-likelihood of cDCC-GARCH model.}
\usage{
cdcc_loglikelihood(param, ht, residuals, stdresids, uncR)
}
\arguments{
\item{param}{cDCC-GARCH parameters(alpha,beta)}

\item{ht}{matrix of conditional variance vectors (T by N)}

\item{residuals}{matrix of residual(de-mean) returns  (T by N)}

\item{stdresids}{matrix of standrdized(De-GARCH) residual returns  (T by N)}

\item{uncR}{unconditional correlation matrix of stdresids  (N by N)}
}
\value{
log-likelihood of cDCC-GARCH model (scaler)
}
\description{
This function calculates log-likelihood of cDCC-GARCH model.
}
