% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_optim.R
\name{dcc_optim}
\alias{dcc_optim}
\title{This function optimizes log-likelihood of DCC-GARCH model.}
\usage{
dcc_optim(param, ht, residuals, stdresids, uncR)
}
\arguments{
\item{param}{DCC-GARCH parameters(alpha,beta)}

\item{ht}{matrix of conditional variance vectors (T by N)}

\item{residuals}{matrix of residual(de-mean) returns  (T by N)}

\item{stdresids}{matrix of standrdized(De-GARCH) residual returns  (T by N)}

\item{uncR}{unconditional correlation matrix of stdresids  (N by N)}
}
\value{
results of optimization
}
\description{
This function optimizes log-likelihood of DCC-GARCH model.
}
