% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_stocks.R
\docType{data}
\name{us_stocks}
\alias{us_stocks}
\title{the closing price data of us stocks in SP500 index from 2006-03-31 to 2014-03-31 from yahoo finance.}
\format{A data frame with 2013 rows and 460 variables:}
\source{
Yahoo finance
}
\description{
the closing price data of us stocks in SP500 index from 2006-03-31 to 2014-03-31 from yahoo finance.
}
